// app/types/enums.ts
enum AloiteState {
  APPROVED = "APPROVED",
  REJECTED = "REJECTED",
  EXPIRED = "EXPIRED",
  NOVOTERS = "NOVOTERS",
  CANCELED = "CANCELED",
  ONGOING = "ONGOING"
}

enum AloiteStep {
  NOT_STARTED = "NOT_STARTED",
  VOTE_COLLECTION = "VOTE_COLLECTION",
  VERIFICATION = "VERIFICATION",
  INITIATION = "INITIATION",
  COMMITTEE_HANDLING = "COMMITTEE_HANDLING",
  FIRST_READING = "FIRST_READING",
  SECOND_READING = "SECOND_READING"
}

export { AloiteState, AloiteStep }